Consolidate snark typed AST codegen
